biodb/requirements.txt
Django>=1.11.3
easy_thumbnails
django-guardian
django-ckeditor
django-tables2
django-addanother
django-simple-history
